modules should be portable from old version
ported the test environment and the builtin module to the new
version of the bot. instead of having a reference to the bot
and dealing with all of the thread safety issues of sending
closures across threads, the bot is single-threaded for now
and takes in a trait object for the module, which it dispatches
events to. a bunch of helpers are provided to assist wtih this